Security fix from Chris Evans: use `unsigned int' for len, not `int'.
1999-02-25  Martin Baulig  <martin@home-of-linux.org>

	* gsu-helper.c (helper_read_password): Security fix from
	Chris Evans: use `unsigned int' for len, not `int'.

+++ mini-utils/gsu/gsu-helper.c
@@ -276,7 +276,7 @@ static char *
 helper_read_password (void)
 {
   char password [BUFSIZ], *text;
-  size_t len;
+  unsigned int len;
 
   if (read (passwd_fd, &len, sizeof (len)) != sizeof (len))
     helper_io_error ("read (passwd_fd)");
